the plan with the move validator is that we seperate out some of the functionality
AttackMap → Only tracks which squares are attacked
MoveValidator → Only validates moves
Board → Only manages game state

moved the piece movement logic back to the pieces lol (a very specific part of it at least)

moved a whole bunch of stuff around. didnt add much of anything new so im not gonna go through it.

i did fix the new game button setting black on the bottom instead of the top. (looked into it and fixed a bunch of toggles on reset)

added delay when FE updates so its not so disorientating